Move filter dropdown to UI
